Added help screen, argument parsing, /bin/sh
Moved rsynccstation to /bin/sh for speed reasons. No bash features were
in use anyhow.

Added basic argument parsing, and a couple of arguments to match. For
exmaple, a username can now be specified where previously that was not
possible.

Added a help screen to describe all of the potential arguments a user
could make use of.